// extensions/vuengine-studio-extension/src/editors/browser/components/Common/Dotting/components/Canvas/types.ts
export interface CanvasGridChangeParams {
  dimensions: {
    columnCount: number;
    rowCount: number;
  };
  indices: {
    topRowIndex: number;
    bottomRowIndex: number;
    leftColumnIndex: number;
    rightColumnIndex: number;
  };
}

export type CanvasGridChangeHandler = (params: CanvasGridChangeParams) => void;